feat(Global): rework to WarheadCore module `mod-bg-reward`
