+++ CMakeLists.txt
@@ -1,6 +1,23 @@
-# Set current dir
-CU_SET_PATH("CUR_DIR_MODULE_BG_REWARD" "${CMAKE_CURRENT_LIST_DIR}")
+#
+# This file is part of the WarheadCore Project. See AUTHORS file for Copyright information
+#
+# This file is free software; as a special exception the author gives
+# unlimited permission to copy and/or distribute it, with or without
+# modifications, as long as this notice is preserved.
+#
+# This program is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY, to the extent permitted by law; without even the
+# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# User has manually chosen to ignore the git-tests, so throw them a warning.
+# This is done EACH compile so they can be alerted about the consequences.
+#
 
-AC_ADD_SCRIPT("${CUR_DIR_MODULE_BG_REWARD}/src/BGReward.cpp")
-AC_ADD_SCRIPT_LOADER("BGReward" "${CUR_DIR_MODULE_BG_REWARD}/src/loader.h")
-CU_ADD_HOOK(AFTER_WORLDSERVER_CMAKE "${CUR_DIR_MODULE_BG_REWARD}/src/cmake/after_ws_install.cmake")
+# Add source files
+AC_ADD_SCRIPT("${CMAKE_CURRENT_LIST_DIR}/src/BGReward_SC.cpp")
+
+# Add scripts to script loader
+AC_ADD_SCRIPT_LOADER("BGReward" "${CMAKE_CURRENT_LIST_DIR}/src/bg_reward_loader.h")
+
+# Add config file
+AC_ADD_CONFIG_FILE("${CMAKE_CURRENT_LIST_DIR}/conf/BGReward.conf.dist")
