+++ src/bg_reward_loader.h
@@ -1,0 +1,30 @@
+/*
+ * This file is part of the WarheadCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef _BG_REWARD_LOADER_H_
+#define _BG_REWARD_LOADER_H_
+
+// From SC
+void AddSC_BGReward();
+
+// Add all
+void AddBGRewardScripts()
+{
+   AddSC_BGReward();
+}
+
+#endif /* _BG_REWARD_LOADER_H_ */
