--- src/loader.h
@@ -1,1 +1,0 @@
-void AddBGRewardScripts();
